+++ src/ui_event/dialog_ui.py
@@ -304,7 +304,7 @@ def save_data(self):
     if write_minio_config_to_file(minio_config=spider_config):
         QMessageBox(QMessageBox.Information, "保存", "配置写入成功,程序即将退出,请重新启动应用以使配置生效！").exec_()
         self.hide()
-        sys.exit()
+        # sys.exit()
     else:
         QMessageBox(QMessageBox.Warning, "错误", "发生未知错误,写入配置失败,请检查log信息！").exec_()
     pass