+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - update selenium page_load_strategy to none
 - print selenium start time
 - update block time generate method
+- update sys sleep method
 
 ## v1.1.4-beta.1.240425
 ### 2024/4/25